feat: get lighting amount from nature remo3

+++ server/crates/atmos_server/src/domain/models/lighting.rs
@@ -16,7 +16,7 @@ pub struct GetLigtingSignalsRequest {
 
 /// 現在の部屋の明るさ
 #[derive(PartialEq, Debug, From)]
-pub struct CurrentLightingAmount(f64);
+pub struct CurrentLightingAmount(f32);
 
 /// 明るさの目標値
 #[derive(PartialEq, Debug, From)]
